'''Vision & Control''' Created basic light seek state machine  * Fixed light.py state machine for hitting the red light  * Updated ram.motion.seek.SeekPoint to move the vehicle forward  * Updated IController::getSpeed/setSpeed is now a double  * Added event for "almost hitting light"  * Remove duplicate view for Vision system data

+++ packages/control/include/IController.h
@@ -62,7 +62,7 @@ public:
     /* @{ */
     
     /** Set the current speed, clamped between -5 and 5 */
-    virtual void setSpeed(int speed) = 0;
+    virtual void setSpeed(double speed) = 0;
 
     /** Sets the current heading in degrees off north */
 //    virtual void setHeading(double degrees) = 0;
@@ -71,7 +71,7 @@ public:
     virtual void setDepth(double depth) = 0;
 
     /** Gets the current speed, a value between -5 and 5 */
-    virtual int getSpeed() = 0;
+    virtual double getSpeed() = 0;
 
     /** The current heading of th*/
 //    virtual double getHeading() = 0;
